refactor(redhat): replace redhat with redhat2
