+++ oval/redhat/redhat.go
@@ -1,6 +1,7 @@
 package redhat
 
 import (
+	"bufio"
 	"bytes"
 	"compress/bzip2"
 	"encoding/xml"
@@ -8,29 +9,33 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"path"
 	"path/filepath"
 	"strings"
 
+	"github.com/cheggaaa/pb"
 	"github.com/spf13/afero"
 	"golang.org/x/xerrors"
-	"gopkg.in/cheggaaa/pb.v1"
 
 	"github.com/aquasecurity/vuln-list-update/utils"
 )
 
 const (
-	ovalDir   = "oval"
-	redhatDir = "redhat"
-	urlFormat = "https://www.redhat.com/security/data/oval/v2/RHEL%s/rhel-%s.oval.xml.bz2"
-	retry     = 5
-)
+	ovalDir      = "oval"
+	redhatDir    = "redhat"
+	urlFormat    = "https://www.redhat.com/security/data/oval/v2/%s"
+	retry        = 5
+	pulpManifest = "PULP_MANIFEST"
 
-var (
-	releases = []string{"6", "7", "8"}
+	testsDir       = "tests"
+	objectsDir     = "objects"
+	statesDir      = "states"
+	definitionsDir = "definitions"
 )
 
 var (
 	ErrInvalidRHSAFormat = errors.New("invalid RHSA-ID format")
+	ErrInvalidCVEFormat  = errors.New("invalid CVE-ID format")
 )
 
 type Config struct {
@@ -50,69 +55,162 @@ func NewConfig() Config {
 }
 
 func (c Config) Update() error {
-	log.Println("Fetching Red Hat OVAL data...")
-	for _, release := range releases {
-		if err := c.update(release); err != nil {
-			return err
+	dirPath := filepath.Join(c.VulnListDir, ovalDir, redhatDir)
+	log.Printf("Remove Red Hat OVAL v2 directory %s", dirPath)
+	if err := os.RemoveAll(dirPath); err != nil {
+		return xerrors.Errorf("failed to remove Red Hat OVAL v2 directory: %w", err)
+	}
+
+	log.Println("Fetching Red Hat OVAL v2 data...")
+	filePaths, err := c.fetchOvalFilePaths()
+	if err != nil {
+		return xerrors.Errorf("failed to get oval file paths: %w", err)
+	}
+	for _, ovalFilePath := range filePaths {
+		log.Printf("Fetching %s", ovalFilePath)
+		if err := c.update(ovalFilePath); err != nil {
+			return xerrors.Errorf("failed to update Red Hat OVAL v2 json: %w", err)
 		}
 	}
+
 	return nil
 }
 
-func (c Config) update(release string) error {
-	url := fmt.Sprintf(c.URLFormat, release, release)
+func (c Config) update(ovalFile string) error {
+	// e.g. RHEL8/storage-gluster-3-including-unpatched.oval.xml.bz2
+	if !strings.HasPrefix(ovalFile, "RHEL") {
+		log.Printf("Skip %s", ovalFile)
+		return nil
+	}
+
+	// e.g. RHEL8/storage-gluster-3-including-unpatched.oval.xml.bz2
+	// => RHEL8/, storage-gluster-3-including-unpatched.oval.xml.bz2
+	dir, file := path.Split(ovalFile)
+	release := strings.TrimPrefix(path.Clean(dir), "RHEL")
+
+	url := fmt.Sprintf(c.URLFormat, ovalFile)
 	res, err := utils.FetchURL(url, "", c.Retry)
 	if err != nil {
-		return xerrors.Errorf("failed to fetch Red Hat OVAL: %w", err)
+		return xerrors.Errorf("failed to fetch Red Hat OVAL v2: %w", err)
 	}
 
 	bzr := bzip2.NewReader(bytes.NewBuffer(res))
-
-	log.Printf("Updating Red Hat %s OVAL data...\n", release)
-	ovalroot := Root{}
-	if err = xml.NewDecoder(bzr).Decode(&ovalroot); err != nil {
-		return xerrors.Errorf("failed to unmarshal Red Hat OVAL XML: %w", err)
+	var ovalroot OvalDefinitions
+	if err := xml.NewDecoder(bzr).Decode(&ovalroot); err != nil {
+		return xerrors.Errorf("failed to unmarshal Red Hat OVAL v2 XML: %w", err)
 	}
 
-	dir := filepath.Join(ovalDir, redhatDir, release)
-	bar := pb.StartNew(len(ovalroot.Definitions.Definitions))
-	for _, def := range ovalroot.Definitions.Definitions {
-		titles := strings.Fields(def.Title)
-		title := strings.Trim(titles[0], ":")
-		if err = c.saveRHSAPerYear(dir, title, def); err != nil {
-			switch err {
-			case ErrInvalidRHSAFormat:
-				continue
-			default:
-				return xerrors.Errorf("unable to save RHEL advisory: %w", err)
+	// e.g. storage-gluster-3-including-unpatched
+	platform := strings.TrimSuffix(file, ".oval.xml.bz2")
+	dirPath := filepath.Join(c.VulnListDir, ovalDir, redhatDir, release, platform)
+
+	// write tests/tests.json file
+	if err := c.writeJSON(filepath.Join(dirPath, testsDir), "tests.json", ovalroot.Tests); err != nil {
+		return xerrors.Errorf("failed to write tests: %w", err)
+	}
+
+	// write objects/objects.json file
+	if err := c.writeJSON(filepath.Join(dirPath, objectsDir), "objects.json", ovalroot.Objects); err != nil {
+		return xerrors.Errorf("failed to write objects: %w", err)
+	}
+
+	// write states/states.json file
+	if err := c.writeJSON(filepath.Join(dirPath, statesDir), "states.json", ovalroot.States); err != nil {
+		return xerrors.Errorf("failed to write states: %w", err)
+	}
+
+	// write definitions
+	bar := pb.StartNew(len(ovalroot.Definitions.Definition))
+	for _, def := range ovalroot.Definitions.Definition {
+		if len(def.Metadata.References) == 0 {
+			continue
+		}
+
+		// RHSA-ID or CVE-ID
+		vulnID := def.Metadata.References[0].RefID
+		for _, ref := range def.Metadata.References {
+			if strings.HasPrefix(ref.RefID, "RHSA-") {
+				vulnID = ref.RefID
 			}
 		}
+
+		if err := c.saveAdvisoryPerYear(filepath.Join(dirPath, definitionsDir), vulnID, def); err != nil {
+			return xerrors.Errorf("failed to save advisory per year: %w", err)
+		}
+
 		bar.Increment()
 	}
 	bar.Finish()
+
 	return nil
 }
 
-func (c Config) saveRHSAPerYear(dirName string, rhsaID string, data interface{}) error {
-	// e.g. RHSA-2018:0094
-	s := strings.Split(rhsaID, ":")
-	if len(s) != 2 {
-		log.Printf("invalid RHSA-ID format: %s\n", rhsaID)
-		return ErrInvalidRHSAFormat
-	}
-	s = strings.Split(s[0], "-")
-	if len(s) != 2 {
-		log.Printf("invalid RHSA-ID format: %s\n", rhsaID)
-		return ErrInvalidRHSAFormat
+func (c Config) fetchOvalFilePaths() ([]string, error) {
+	res, err := utils.FetchURL(fmt.Sprintf(c.URLFormat, pulpManifest), "", c.Retry)
+	if err != nil {
+		return nil, xerrors.Errorf("failed to fetch PULP_MANIFEST: %w", err)
 	}
 
-	yearDir := filepath.Join(c.VulnListDir, dirName, s[1])
-	if err := c.AppFs.MkdirAll(yearDir, os.ModePerm); err != nil {
+	var ovalFilePaths []string
+	scanner := bufio.NewScanner(bytes.NewReader(res))
+	for scanner.Scan() {
+		ss := strings.Split(scanner.Text(), ",")
+		if len(ss) < 3 {
+			return nil, xerrors.Errorf("failed to parse PULP_MANIFEST: %w", err)
+		}
+		if !strings.Contains(ss[0], "including-unpatched") {
+			continue
+		}
+
+		ovalFilePaths = append(ovalFilePaths, ss[0])
+	}
+	return ovalFilePaths, nil
+}
+
+func (c Config) saveAdvisoryPerYear(dirName string, id string, def Definition) error {
+	var year string
+	if strings.HasPrefix(id, "CVE") {
+		s := strings.Split(id, "-")
+		if len(s) != 3 {
+			log.Printf("invalid CVE-ID format: %s\n", id)
+			return ErrInvalidCVEFormat
+		}
+		year = s[1]
+	} else {
+		// e.g. RHSA-2018:0094
+		s := strings.Split(id, ":")
+		if len(s) != 2 {
+			log.Printf("invalid RHSA-ID format: %s\n", id)
+			return ErrInvalidRHSAFormat
+		}
+		s = strings.Split(s[0], "-")
+		if len(s) != 2 {
+			log.Printf("invalid RHSA-ID format: %s\n", id)
+			return ErrInvalidRHSAFormat
+		}
+		year = s[1]
+	}
+
+	fileFmt := "%s.json"
+	if strings.HasPrefix(def.ID, "oval:com.redhat.unaffected:def") {
+		fileFmt = "%s.unaffected.json"
+	}
+
+	yearDir := filepath.Join(dirName, year)
+	if err := c.writeJSON(yearDir, fmt.Sprintf(fileFmt, id), def); err != nil {
+		return xerrors.Errorf("unable to write a JSON file: %w", err)
+	}
+	return nil
+}
+
+func (c Config) writeJSON(dirName, fileName string, data interface{}) error {
+	if err := c.AppFs.MkdirAll(dirName, os.ModePerm); err != nil {
 		return xerrors.Errorf("failed to create a year dir: %w", err)
 	}
 
-	filePath := filepath.Join(yearDir, fmt.Sprintf("%s.json", rhsaID))
 	fs := utils.NewFs(c.AppFs)
+
+	filePath := filepath.Join(dirName, fileName)
 	if err := fs.WriteJSON(filePath, data); err != nil {
 		return xerrors.Errorf("failed to write file: %w", err)
 	}
