+++ oval/redhat/testdata/ansible-2-including-unpatched.oval.xml.bz2
