--- oval/redhat/testdata/golden/RHBA-2011:1656.json
@@ -1,168 +1,0 @@
-{
-  "XMLName": {
-    "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-    "Local": "definition"
-  },
-  "ID": "oval:com.redhat.rhba:def:20111656",
-  "Class": "patch",
-  "Title": "RHBA-2011:1656: mod_nss bug fix update  (None)",
-  "Affecteds": [
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "affected"
-      },
-      "Family": "unix",
-      "Platforms": [
-        "Red Hat Enterprise Linux 6"
-      ]
-    }
-  ],
-  "References": [
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "reference"
-      },
-      "Source": "RHBA",
-      "RefID": "RHBA-2011:1656",
-      "RefURL": "https://access.redhat.com/errata/RHBA-2011:1656"
-    },
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "reference"
-      },
-      "Source": "CVE",
-      "RefID": "CVE-2011-4973",
-      "RefURL": "https://access.redhat.com/security/cve/CVE-2011-4973"
-    }
-  ],
-  "Description": "The mod_nss module provides strong cryptography for the Apache HTTP Server via the Secure Sockets Layer (SSL) and Transport Layer Security (TLS) protocols, using the Network Security Services (NSS) security library.\n\nThis update fixes the following bugs: \n\n* When the NSS library was not initialized and mod_nss tried to clear its SSL cache on start-up, mod_nss terminated unexpectedly when the NSS library was built with debugging enabled. With this update, mod_nss does not try to clear the SSL cache in the described scenario, thus preventing this bug. (BZ#691502)\n\n* Previously, a static array containing the arguments for launching the nss_pcache command was overflowing the size by one. This could lead to a variety of issues including unexpected termination. This bug has been fixed, and mod_nss now uses properly sized static array when launching nss_pcache. (BZ#714154)\n\n* Prior to this update, client certificates were only retrieved during the initial SSL handshake if the NSSVerifyClient option was set to \"require\" or \"optional\". Also, the FakeBasicAuth option only retrieved Common Name rather than the entire certificate subject. Consequently, it was possible to spoof an identity using that option. This bug has been fixed, the FakeBasicAuth option is now prefixed with \"/\" and is thus compatible with OpenSSL, and certificates are now retrieved on all subsequent requests beyond the first one. (BZ#702437)\n\nUsers of mod_nss are advised to upgrade to this updated package, which fixes these bugs.",
-  "Advisory": {
-    "XMLName": {
-      "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-      "Local": "advisory"
-    },
-    "Severity": "None",
-    "Cves": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "cve"
-        },
-        "CveID": "CVE-2011-4973",
-        "Cvss2": "4.0/AV:N/AC:H/Au:N/C:P/I:P/A:N",
-        "Cvss3": "",
-        "Cwe": "",
-        "Impact": "moderate",
-        "Href": "https://access.redhat.com/security/cve/CVE-2011-4973",
-        "Public": "20110504"
-      }
-    ],
-    "Bugzillas": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "bugzilla"
-        },
-        "ID": "702437",
-        "URL": "https://bugzilla.redhat.com/702437",
-        "Title": "may be possible to spoof mod_nss FakeBasicAuth"
-      },
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "bugzilla"
-        },
-        "ID": "714154",
-        "URL": "https://bugzilla.redhat.com/714154",
-        "Title": "overrunning array when executing nss_pcache"
-      }
-    ],
-    "AffectedCPEList": [
-      "cpe:/o:redhat:enterprise_linux:6"
-    ],
-    "Issued": {
-      "Date": "2011-12-05"
-    },
-    "Updated": {
-      "Date": "2011-12-06"
-    }
-  },
-  "Criteria": {
-    "XMLName": {
-      "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-      "Local": "criteria"
-    },
-    "Operator": "AND",
-    "Criterias": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "criteria"
-        },
-        "Operator": "OR",
-        "Criterias": null,
-        "Criterions": [
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criterion"
-            },
-            "Negate": false,
-            "TestRef": "oval:com.redhat.rhba:tst:20111656003",
-            "Comment": "Red Hat Enterprise Linux 6 Client is installed"
-          },
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criterion"
-            },
-            "Negate": false,
-            "TestRef": "oval:com.redhat.rhba:tst:20111656004",
-            "Comment": "Red Hat Enterprise Linux 6 Server is installed"
-          },
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criterion"
-            },
-            "Negate": false,
-            "TestRef": "oval:com.redhat.rhba:tst:20111656005",
-            "Comment": "Red Hat Enterprise Linux 6 Workstation is installed"
-          },
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criterion"
-            },
-            "Negate": false,
-            "TestRef": "oval:com.redhat.rhba:tst:20111656006",
-            "Comment": "Red Hat Enterprise Linux 6 ComputeNode is installed"
-          }
-        ]
-      }
-    ],
-    "Criterions": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "criterion"
-        },
-        "Negate": false,
-        "TestRef": "oval:com.redhat.rhba:tst:20111656001",
-        "Comment": "mod_nss is earlier than 0:1.0.8-13.el6"
-      },
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "criterion"
-        },
-        "Negate": false,
-        "TestRef": "oval:com.redhat.rhba:tst:20111656002",
-        "Comment": "mod_nss is signed with Red Hat redhatrelease2 key"
-      }
-    ]
-  }
-}
