--- oval/redhat/testdata/golden/RHSA-2019:0968.json
@@ -1,170 +1,0 @@
-{
-  "XMLName": {
-    "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-    "Local": "definition"
-  },
-  "ID": "oval:com.redhat.rhsa:def:20190968",
-  "Class": "patch",
-  "Title": "RHSA-2019:0968: edk2 security update (Important)",
-  "Affecteds": [
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "affected"
-      },
-      "Family": "unix",
-      "Platforms": [
-        "Red Hat Enterprise Linux 8"
-      ]
-    }
-  ],
-  "References": [
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "reference"
-      },
-      "Source": "RHSA",
-      "RefID": "RHSA-2019:0968",
-      "RefURL": "https://access.redhat.com/errata/RHSA-2019:0968"
-    },
-    {
-      "XMLName": {
-        "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-        "Local": "reference"
-      },
-      "Source": "CVE",
-      "RefID": "CVE-2018-12180",
-      "RefURL": "https://access.redhat.com/security/cve/CVE-2018-12180"
-    }
-  ],
-  "Description": "EDK (Embedded Development Kit) is a project to enable UEFI support for Virtual Machines. This package contains a sample 64-bit UEFI firmware for QEMU and KVM.\n\nSecurity Fix(es):\n\n* edk2: Buffer Overflow in BlockIo service for RAM disk (CVE-2018-12180)\n\nFor more details about the security issue(s), including the impact, a CVSS score, acknowledgments, and other related information, refer to the CVE page(s) listed in the References section.",
-  "Advisory": {
-    "XMLName": {
-      "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-      "Local": "advisory"
-    },
-    "Severity": "Important",
-    "Cves": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "cve"
-        },
-        "CveID": "CVE-2018-12180",
-        "Cvss2": "",
-        "Cvss3": "8.3/CVSS:3.0/AV:N/AC:L/PR:N/UI:N/S:C/C:L/I:L/A:L",
-        "Cwe": "CWE-125 CWE-787",
-        "Impact": "",
-        "Href": "https://access.redhat.com/security/cve/CVE-2018-12180",
-        "Public": "20190226"
-      }
-    ],
-    "Bugzillas": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "bugzilla"
-        },
-        "ID": "1683372",
-        "URL": "https://bugzilla.redhat.com/1683372",
-        "Title": "CVE-2018-12180 edk2: Buffer Overflow in BlockIo service for RAM disk"
-      }
-    ],
-    "AffectedCPEList": [
-      "cpe:/a:redhat:enterprise_linux:8"
-    ],
-    "Issued": {
-      "Date": "2019-05-07"
-    },
-    "Updated": {
-      "Date": "2019-05-07"
-    }
-  },
-  "Criteria": {
-    "XMLName": {
-      "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-      "Local": "criteria"
-    },
-    "Operator": "AND",
-    "Criterias": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "criteria"
-        },
-        "Operator": "OR",
-        "Criterias": [
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criteria"
-            },
-            "Operator": "AND",
-            "Criterias": null,
-            "Criterions": [
-              {
-                "XMLName": {
-                  "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-                  "Local": "criterion"
-                },
-                "Negate": false,
-                "TestRef": "oval:com.redhat.rhsa:tst:20190968001",
-                "Comment": "edk2-ovmf is earlier than 0:20180508gitee3198e672e2-9.el8_0.1"
-              },
-              {
-                "XMLName": {
-                  "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-                  "Local": "criterion"
-                },
-                "Negate": false,
-                "TestRef": "oval:com.redhat.rhsa:tst:20190968002",
-                "Comment": "edk2-ovmf is signed with Red Hat redhatrelease2 key"
-              }
-            ]
-          },
-          {
-            "XMLName": {
-              "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-              "Local": "criteria"
-            },
-            "Operator": "AND",
-            "Criterias": null,
-            "Criterions": [
-              {
-                "XMLName": {
-                  "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-                  "Local": "criterion"
-                },
-                "Negate": false,
-                "TestRef": "oval:com.redhat.rhsa:tst:20190968003",
-                "Comment": "edk2-aarch64 is earlier than 0:20180508gitee3198e672e2-9.el8_0.1"
-              },
-              {
-                "XMLName": {
-                  "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-                  "Local": "criterion"
-                },
-                "Negate": false,
-                "TestRef": "oval:com.redhat.rhsa:tst:20190968004",
-                "Comment": "edk2-aarch64 is signed with Red Hat redhatrelease2 key"
-              }
-            ]
-          }
-        ],
-        "Criterions": null
-      }
-    ],
-    "Criterions": [
-      {
-        "XMLName": {
-          "Space": "http://oval.mitre.org/XMLSchema/oval-definitions-5",
-          "Local": "criterion"
-        },
-        "Negate": false,
-        "TestRef": "oval:com.redhat.rhsa:tst:20190966005",
-        "Comment": "Red Hat Enterprise Linux 8 is installed"
-      }
-    ]
-  }
-}
