+++ oval/redhat/testdata/golden/ansible-2-including-unpatched/CVE-2020-1734.json
