+++ oval/redhat/testdata/golden/ansible-2-including-unpatched/RHSA-2019-3927.json
