+++ oval/redhat/testdata/golden/ansible-2-including-unpatched/RHSA-2020-0215.json
