+++ oval/redhat/testdata/golden/ansible-2-including-unpatched/states.json
