+++ oval/redhat/testdata/golden/dotnet-3.1-including-unpatched/CVE-2020-0605.unaffected.json
