+++ oval/redhat/testdata/golden/dotnet-3.1-including-unpatched/RHSA-2020-2249.json
