+++ oval/redhat/testdata/golden/dotnet-3.1-including-unpatched/states.json
