+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/CVE-2014-3209.json
