+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/CVE-2016-5361.json
