+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/CVE-2018-5389.json
