+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/CVE-2020-28935.json
