+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/RHBA-2014-1396.json
