+++ oval/redhat/testdata/golden/rhel-6-extras-including-unpatched/tests.json
