+++ oval/redhat/testdata/rhel-6-extras-including-unpatched.oval.xml.bz2
