--- oval/redhat2/testdata/test.txt
@@ -1,1 +1,0 @@
-test
